Refactored the views into partial layouts, added the Users controller and made major CSS styling improvements.
Moved several portions of the application's main layout into partials,
such as the stylesheet definitions, the header and the footer. Removed
the default index.html and modified the routes to match corresponding
urls throughout the site. Added a home and legal static view. Generated
the Users controller with an action of 'new' in order to implement the
join page (at least in terms of routing). Styled the entire site with
CSS declarations and added a logo image to the main layout. Last, but
not least, I generated integration tests for the layout links and filled
in various test definitions to test the various routes.